# internal/server/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "suxen control-plane API",
    "version": "dev",
    "description": "Administrative API for repositories, access control, retention, provenance, and operations. OCI Distribution endpoints under /v2 and /repository/{name}/v2 use the OCI error envelope and are intentionally outside this control-plane contract."
  },
  "servers": [
    {
      "url": "/",
      "description": "Current suxen server"
    }
  ],
  "security": [
    {},
    {
      "basicAuth": []
    },
    {
      "bearerAuth": []
    },
    {
      "oidcSession": []
    }
  ],
...
  "paths": {
...
    "/api/v1/provision": {
      "get": {
        "operationId": "listManagedResources",
        "x-suxen-required-privilege": "admin:provision:read",
        "summary": "List declaratively managed resources",
        "description": "Lists declarative ownership metadata without exposing comparison hashes or secret fingerprints.",
        "tags": ["Operations"],
        "parameters": [
          {"$ref": "#/components/parameters/Limit"},
          {"$ref": "#/components/parameters/Page"},
          {"$ref": "#/components/parameters/Cursor"}
        ],
        "responses": {
          "200": {
            "description": "Managed resource page",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/ManagedResourcePage"}
              }
            }
          },
          "400": {"$ref": "#/components/responses/BadRequest"},
          "409": {"$ref": "#/components/responses/Conflict"},
          "401": {"$ref": "#/components/responses/Unauthorized"},
          "403": {"$ref": "#/components/responses/Forbidden"},
          "default": {"$ref": "#/components/responses/Error"}
        }
      },
      "post": {
        "operationId": "applyProvisioningDocument",
        "x-suxen-required-privilege": "admin:provision:write",
        "summary": "Apply declarative control-plane resources",
        "description": "Reconciles a complete desired-state document. Requires the admin:provision:write privilege. Secret references are resolved by the server and secret values are never returned.",
        "tags": ["Operations"],
        "parameters": [
          {"$ref": "#/components/parameters/ProvisionDryRun"},
          {"$ref": "#/components/parameters/ProvisionPrune"}
        ],
        "requestBody": {"$ref": "#/components/requestBodies/ProvisioningDocument"},
        "responses": {
          "200": {
            "description": "Deterministic reconciliation report",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/ProvisioningReport"}
              }
            }
          },
          "400": {"$ref": "#/components/responses/BadRequest"},
          "401": {"$ref": "#/components/responses/Unauthorized"},
          "403": {"$ref": "#/components/responses/Forbidden"},
          "415": {"$ref": "#/components/responses/UnsupportedMediaType"},
          "default": {"$ref": "#/components/responses/Error"}
        }
      }
    },
...
    "/api/v1/blob-stores/{name}": {
      "parameters": [{"$ref": "#/components/parameters/BlobStoreName"}],
      "get": {
        "operationId": "getBlobStore",
        "x-suxen-required-privilege": "admin:blob-stores:read",
        "summary": "Get a blob store",
        "tags": ["Blob stores"],
        "responses": {
          "200": {
            "description": "Blob store",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/BlobStore"}
              }
            }
          },
          "401": {"$ref": "#/components/responses/Unauthorized"},
          "403": {"$ref": "#/components/responses/Forbidden"},
          "404": {"$ref": "#/components/responses/NotFound"},
          "default": {"$ref": "#/components/responses/Error"}
        }
      },
      "put": {
        "operationId": "updateBlobStore",
        "x-suxen-required-privilege": "admin:blob-stores:write",
        "parameters": [{"$ref": "#/components/parameters/Force"}],
        "summary": "Update blob store operational attributes",
        "description": "The driver, configuration reference and physical destination remain fixed for the resource's lifetime, including the default store. Only operational attributes can change. Credential rotation behind the same reference may require a process restart; changing destination requires a new store and drain.",
        "tags": ["Blob stores"],
        "requestBody": {"$ref": "#/components/requestBodies/BlobStoreUpdate"},
        "responses": {
          "200": {
            "description": "Blob store updated",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/BlobStore"}
              }
            }
          },
          "201": {
            "description": "Blob store created",
            "headers": {"Location": {"$ref": "#/components/headers/Location"}},
            "content": {"application/json": {"schema": {"$ref": "#/components/schemas/BlobStore"}}}
          },
          "400": {"$ref": "#/components/responses/BadRequest"},
          "401": {"$ref": "#/components/responses/Unauthorized"},
          "403": {"$ref": "#/components/responses/Forbidden"},
          "404": {"$ref": "#/components/responses/NotFound"},
          "409": {"$ref": "#/components/responses/Conflict"},
          "default": {"$ref": "#/components/responses/Error"}
        }
      },
      "delete": {
        "operationId": "deleteBlobStore",
        "x-suxen-required-privilege": "admin:blob-stores:write",
        "parameters": [{"$ref": "#/components/parameters/Force"}],
        "summary": "Delete a blob store",
        "tags": ["Blob stores"],
        "responses": {
          "204": {"description": "Blob store deleted"},
          "400": {"$ref": "#/components/responses/BadRequest"},
          "401": {"$ref": "#/components/responses/Unauthorized"},
          "403": {"$ref": "#/components/responses/Forbidden"},
          "404": {"$ref": "#/components/responses/NotFound"},
          "409": {"$ref": "#/components/responses/Conflict"},
          "default": {"$ref": "#/components/responses/Error"}
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "basicAuth": {
        "type": "http",
        "scheme": "basic",
        "description": "Local username and password."
      },
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "API token or OIDC JWT",
        "description": "Suxen API token or a configured OIDC bearer token. API-token scopes use repository:<name|*>:<read|write|delete|annotate|manage> and admin:<resource|*>:<action|*>. An empty scope list does not further restrict the user's roles.",
        "x-suxen-scope-syntax": [
          "repository:<name|*>:<read|write|delete|annotate|manage>",
          "admin:<resource|*>:<action|*>"
        ]
      },
      "oidcSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "suxen_oidc_token",
        "description": "HTTP-only browser session established by the configured OIDC login flow."
      }
    },
    "headers": {
      "Location": {
        "description": "Canonical path of the created resource.",
        "schema": {"type": "string"}
      }
    },
    "parameters": {
      "BlobStoreName": {
        "name": "name",
        "in": "path",
        "required": true,
        "description": "Blob store name",
        "schema": {"$ref": "#/components/schemas/ResourceName"}
      },
...
      "Limit": {
        "name": "limit",
        "in": "query",
        "required": false,
        "description": "Maximum number of records",
        "schema": {
          "type": "integer",
          "minimum": 1,
          "maximum": 200,
          "default": 100
        }
      },
      "Page": {
        "name": "page",
        "in": "query",
        "required": false,
        "description": "1-based random-access page. Cannot be combined with cursor; pages do not enforce a collection snapshot.",
        "schema": {
          "type": "integer",
          "minimum": 1
        }
      },
      "Cursor": {
        "name": "cursor",
        "in": "query",
        "required": false,
        "description": "Opaque continuation cursor returned by the same operation and filters",
        "schema": {
          "type": "string",
          "maxLength": 2048
        }
      },
...
      "ProvisionDryRun": {
        "name": "dryRun",
        "in": "query",
        "required": false,
        "description": "Validate and report reconciliation decisions without mutating resources",
        "schema": {
          "type": "boolean",
          "default": false
        }
      },
      "ProvisionPrune": {
        "name": "prune",
        "in": "query",
        "required": false,
        "description": "Delete provision-owned resources omitted from the desired state after reference validation",
        "schema": {
          "type": "boolean",
          "default": false
        }
      },
      "Force": {
        "name": "force",
        "in": "query",
        "required": false,
        "description": "Transfer a managed resource out of declarative provisioning before applying the mutation",
        "schema": {
          "type": "boolean",
          "default": false
        }
      },
...
    },
    "requestBodies": {
      "ProvisioningDocument": {
        "required": true,
        "content": {
          "application/json": {
            "schema": {"$ref": "#/components/schemas/ProvisioningDocument"}
          },
          "application/yaml": {
            "schema": {"$ref": "#/components/schemas/ProvisioningDocument"}
          }
        }
      },
...
      "BlobStoreUpdate": {
        "required": true,
        "content": {
          "application/json": {
            "schema": {"$ref": "#/components/schemas/BlobStoreUpdateRequest"}
          }
        }
      },
...
    },
    "responses": {
...
      "BadRequest": {
        "description": "Invalid request. Account validation codes include invalid_username and password_required; the reserved repository name default returns reserved_repository_name.",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Authentication required. Control-plane routes answer 401 without a WWW-Authenticate challenge so browsers do not raise the native Basic-auth popup; clients authenticate with a bearer token or the cookie session and treat 401 as a prompt to sign in.",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
      "Forbidden": {
        "description": "Insufficient privilege",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
      "NotFound": {
        "description": "Resource not found",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
      "Conflict": {
        "description": "Resource already exists, is declaratively managed, or is still referenced",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
      "UnsupportedMediaType": {
        "description": "Request body must use a supported media type",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
      "Error": {
        "description": "Error response",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      }
    },
    "schemas": {
...
      "Timestamp": {
        "type": "string",
        "format": "date-time"
      },
...
      "ProvisioningReport": {
        "type": "object",
        "additionalProperties": false,
        "required": ["dryRun", "prune", "results"],
        "properties": {
          "dryRun": {"type": "boolean"},
          "prune": {"type": "boolean"},
          "results": {
            "type": "array",
            "items": {"$ref": "#/components/schemas/ProvisioningResult"}
          }
        }
      },
...
      "ManagedResourcePage": {
        "type": "object",
        "required": ["items"],
        "properties": {
          "items": {
            "type": "array",
            "items": {"$ref": "#/components/schemas/ManagedResource"}
          },
          "nextCursor": {"type": "string"},
          "total": {"type": "integer"}
        }
      },
...
      "BlobStore": {
        "type": "object",
        "required": ["name", "driver", "configurationRef", "managed", "state", "createdAt"],
        "properties": {
          "name": {"$ref": "#/components/schemas/ResourceName"},
          "driver": {
            "type": "string",
            "pattern": "^[a-z0-9_-]+$",
            "example": "fs"
          },
		  "configurationRef": {"$ref": "#/components/schemas/ConfigurationReference"},
		  "attributes": {"$ref": "#/components/schemas/BlobStoreAttributes"},
		  "managed": {"type": "boolean", "readOnly": true},
          "state": {
            "type": "string",
            "enum": ["active", "draining", "drained"],
            "readOnly": true
          },
          "drainTarget": {
            "allOf": [{"$ref": "#/components/schemas/ResourceName"}],
            "readOnly": true
          },
          "createdAt": {
            "allOf": [{"$ref": "#/components/schemas/Timestamp"}],
            "readOnly": true
          }
        }
      },
...
    }
  }
}